TOOL-4/TYPE-2: type-check scripts/ instead of excluding it
The mypy pre-commit hook excluded scripts/, which holds the data converter and
the benchmark tooling - most of the repo's non-library Python. Nothing ran mypy
over it, and its annotations had drifted to 15 errors: two `# type: ignore`
codes mypy no longer emits (so the ignore silenced nothing), two implicit
`Optional` defaults that `no_implicit_optional = true` was already configured to
reject, a dict annotated with a narrower value type than it is assigned, and
four bbox lists annotated `int` while `Boundaries` declares `float`.

All fixed as annotations - no runtime behaviour changes - and the exclude
narrowed so they stay fixed. Two of the four errors mypy reported in
tests/auxiliaries.py, reached by following imports from scripts/, are fixed
alongside: `bool(np.all(...))` where the signature promises `bool`, and passing
PROJECT_ROOT to a `cwd: Path` parameter that stringifies it itself.

test_scripts_are_type_checked_by_the_hook guards the exclude: re-adding
scripts/ to it is a one-word edit that reports nothing, which the two existing
tests over `ignore_errors` cannot see.

+++ scripts/file_converter.py
@@ -103,7 +103,7 @@ def to_numpy_array(values: list[Any], dtype: str) -> NDArray[Any]:
     return np.array(values, dtype=dtype)
 
 
-def to_bbox_vector(values: list[int]) -> BoundaryArray:
+def to_bbox_vector(values: list[float]) -> BoundaryArray:
     return to_numpy_array(values, dtype=DTYPE_FORMAT_SIGNED_I_NUMPY)
 
 
@@ -116,10 +116,10 @@ def convert_bboxes_to_numpy(
     Returns:
         Tuple of numpy arrays (xmax, xmin, ymax, ymin)
     """
-    xmax_list: list[int] = []
-    xmin_list: list[int] = []
-    ymax_list: list[int] = []
-    ymin_list: list[int] = []
+    xmax_list: list[float] = []
+    xmin_list: list[float] = []
+    ymax_list: list[float] = []
+    ymin_list: list[float] = []
     for bounds in bboxes:
         xmax_list.append(bounds.xmax)
         xmin_list.append(bounds.xmin)
